moving RegisterForm out
