+++ src/components/Auth/components/RegisterForm.spec.jsx
