--- src/components/Auth/components/RegisterForm/index.js
@@ -1,3 +1,0 @@
-export { default as PasswordField } from './PasswordField.jsx';
-export { default as RoleField } from './RoleField.jsx';
-export { default as FormChecker } from './FormChecker.jsx';
