+++ src/components/Auth/components/index.js
@@ -1,3 +1,2 @@
-export { default as Auth } from './Auth.jsx';
-export { default as Login } from './Login.jsx';
-export { default as RegisterForm } from './RegisterForm/RegisterForm.jsx';
+export { default as LoginForm } from './LoginForm.jsx';
+export { default as RegisterForm } from './RegisterForm.jsx';
